requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backtrader==1.9.78.123
beautifulsoup4==4.12.3
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
fonttools==4.52.4
frozendict==2.4.4
frozenlist==1.4.1
greenlet==3.0.3
html5lib==1.1
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.1
langchain-core==0.2.1
langchain-nvidia-ai-endpoints==0.0.18
langchain-text-splitters==0.2.0
langgraph==0.0.55
langsmith==0.1.63
lxml==5.2.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
multidict==6.0.5
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pandas==2.2.2
parso==0.8.4
peewee==3.17.5
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.44
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
pyrate-limiter==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.2
requests-cache==1.2.0
requests-ratelimiter==0.6.0
scipy==1.13.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
tenacity==8.3.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
uuid6==2024.1.12
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.9.4
yfinance==0.2.40
